requirements.txt
Django>=4.1.3,<4.2
wagtail>=4.1.1,<4.2

django_extensions==3.2.1
django-recaptcha==3.0.0
wagtail-seo==2.3.0

# wagtail-generic-chooser==0.5

psycopg2==2.9.5
psycopg2-binary==2.9.5
gunicorn==20.1.0
